Создаёт отдельный компонент для комментариев, удаляет лишние компоненты
