+++ src/view/film-details.js
@@ -1,41 +1,13 @@
-import {EMOJIS} from "../const.js";
 import {generateTemplate, checkActiveElement, createElement} from "../util";
 
 const createGenresTemplate = (genre) => {
   return `<span class="film-details__genre">${genre}</span>`;
 };
 
-const createCommentsTemplate = (commentElement) => {
-  const {emoji, comment, commentDate, author} = commentElement;
-
-  return `<li class="film-details__comment">
-    <span class="film-details__comment-emoji">
-      <img src="./images/emoji/${emoji}.png" width="55" height="55" alt="emoji-smile">
-    </span>
-    <div>
-      <p class="film-details__comment-text">${comment}</p>
-      <p class="film-details__comment-info">
-        <span class="film-details__comment-author">${author}</span>
-        <span class="film-details__comment-day">${commentDate}</span>
-        <button class="film-details__comment-delete">Delete</button>
-      </p>
-    </div>
-  </li>`;
-};
-
-const createEmojiTemplate = (emoji) => {
-  return `<input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-${emoji}" value="${emoji} checked">
-    <label class="film-details__emoji-label" for="emoji-${emoji}">
-      <img src="./images/emoji/${emoji}.png" width="30" height="30" alt="emoji-${emoji}">
-    </label>`;
-};
-
 const createFilmDetailsTemplate = (film) => {
-  const {filmPoster, filmTitle, rating, filmDuration, genres, description, filmOriginTitle, director, writers, actors, country, ageLimit, isWatchlist, isWatched, isFavorite, comments, date} = film;
+  const {filmPoster, filmTitle, rating, filmDuration, genres, description, filmOriginTitle, director, writers, actors, country, ageLimit, isWatchlist, isWatched, isFavorite, date} = film;
 
-  const commentTemplate = generateTemplate(comments, createCommentsTemplate);
   const filmGenresTemplate = generateTemplate(genres, createGenresTemplate);
-  const emojiTemplate = generateTemplate(EMOJIS, createEmojiTemplate);
 
   const releaseDate = date.toLocaleString(`en-GB`, {year: `numeric`, month: `long`, day: `numeric`});
 
@@ -120,27 +92,6 @@ const createFilmDetailsTemplate = (film) => {
           </section>
         </div>
 
-        <div class="form-details__bottom-container">
-          <section class="film-details__comments-wrap">
-            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-            <ul class="film-details__comments-list">
-              ${commentTemplate}
-            </ul>
-
-            <div class="film-details__new-comment">
-              <div for="add-emoji" class="film-details__add-emoji-label"></div>
-
-              <label class="film-details__comment-label">
-                <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-              </label>
-
-              <div class="film-details__emoji-list">
-                ${emojiTemplate}
-              </div>
-            </div>
-          </section>
-        </div>
       </form>
     </section>`
   );
