+++ src/view/films-block.js
@@ -2,11 +2,20 @@ import {createElement} from "../util.js";
 
 const createFilmsBlockTemplate = () => {
   return (
-    `<section class="films"></section>`
+    `<section class="films">
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+
+        <div class="films-list__container">
+
+        </div>
+
+      </section>
+    </section>`
   );
 };
 
-export default class FilmsList {
+export default class FilmsBlock {
   constructor() {
     this._element = null;
   }
